[KhanhNHQ] Applying migration '20220910173845_Krowd_v.1.0.3', done 274

+++ RevenueSharingInvest.API/Controllers/ProjectEntityController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using RevenueSharingInvest.Business.Models.Constant;
 using RevenueSharingInvest.Business.Services;
 using RevenueSharingInvest.Data.Models.DTOs;
 using RevenueSharingInvest.Data.Models.Entities;
@@ -9,6 +10,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RevenueSharingInvest.API.Controllers
@@ -32,46 +34,82 @@ namespace RevenueSharingInvest.API.Controllers
 
         //CREATE
         [HttpPost]
-        //[Authorize(Roles = "PROJECT_MANAGER")]
+        [Authorize(Roles = "PROJECT_MANAGER")]
         public async Task<IActionResult> CreateProjectEntity([FromBody] CreateUpdateProjectEntityDTO projectEntityDTO)
         {
-            //ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
 
             //PROJECT_MANAGER
-            //if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
-            //{
+            if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
+            {
                 var result = await _projectEntityService.CreateProjectEntity(projectEntityDTO, null);
                 return Ok(result);
-            //}
-            //return StatusCode((int)HttpStatusCode.Forbidden, "Only user with role PROJECT_MANAGER can perform this action!!!");
         }
+            return StatusCode((int) HttpStatusCode.Forbidden, "Only user with role PROJECT_MANAGER can perform this action!!!");
+    }
 
+        //GET ALL
         [HttpGet]
-        [Authorize(Roles = "PROJECT_MANAGER")]
-        public async Task<IActionResult> GetAllProjectEntities(int pageIndex, int pageSize)
+        [Route("{project_id}")]
+        [AllowAnonymous]
+        public async Task<IActionResult> GetProjectEntityByProjectIdAndType(Guid project_Id, string type)
         {
-            var result = new List<GetProjectEntityDTO>();
-            result = await _projectEntityService.GetAllProjectEntities(pageIndex, pageSize);
-            return Ok(result);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+
+            //ALL ROLES
+            if (currentUser.roleId.Equals(currentUser.adminRoleId)
+                || currentUser.roleId.Equals(currentUser.businessManagerRoleId)
+                || currentUser.roleId.Equals(currentUser.projectManagerRoleId)
+                || currentUser.roleId.Equals(currentUser.investorRoleId)
+                || currentUser.roleId.Equals(""))
+            {
+                var result = new List<GetProjectEntityDTO>();
+                result = await _projectEntityService.GetProjectEntityByProjectIdAndType(project_Id, type, currentUser);
+                return Ok(result);
+            }
+
+            return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
         }
 
+        //GET BY ID
         [HttpGet]
         [Route("{id}")]
-        [Authorize(Roles = "PROJECT_MANAGER")]
+        [AllowAnonymous]
         public async Task<IActionResult> GetProjectEntityById(Guid id)
         {
-            GetProjectEntityDTO dto = new GetProjectEntityDTO();
-            dto = await _projectEntityService.GetProjectEntityById(id);
-            return Ok(dto);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+
+            //ALL ROLES
+            if (currentUser.roleId.Equals(currentUser.adminRoleId)
+                || currentUser.roleId.Equals(currentUser.businessManagerRoleId)
+                || currentUser.roleId.Equals(currentUser.projectManagerRoleId)
+                || currentUser.roleId.Equals(currentUser.investorRoleId)
+                || currentUser.roleId.Equals(""))
+            {
+                GetProjectEntityDTO dto = new GetProjectEntityDTO();
+                dto = await _projectEntityService.GetProjectEntityById(id, currentUser);
+                return Ok(dto);
+            }
+
+            return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
         }
 
+
+        //UPDATE
         [HttpPut]
         [Route("{id}")]
         [Authorize(Roles = "PROJECT_MANAGER")]
         public async Task<IActionResult> UpdateProjectEntity([FromForm] CreateUpdateProjectEntityDTO projectEntityDTO, Guid id)
         {
-            var result = await _projectEntityService.UpdateProjectEntity(projectEntityDTO, id);
-            return Ok(result);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+
+            //PROJECT_MANAGER
+            if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
+            {
+                var result = await _projectEntityService.UpdateProjectEntity(projectEntityDTO, id, currentUser);
+                return Ok(result);
+            }
+            return StatusCode((int)HttpStatusCode.Forbidden, "Only user with role PROJECT_MANAGER can perform this action!!!");
         }
 
         //[HttpPut]
@@ -81,20 +119,38 @@ namespace RevenueSharingInvest.API.Controllers
         //    return Ok(result);
         //}
 
+        //UPDATE PRIORITY
         [HttpPut]
+        [Route("priority")]
+        [Authorize(Roles = "PROJECT_MANAGER")]
         public async Task<IActionResult> UpdateProjectEntityPriority([FromBody] List<ProjectEntityUpdateDTO> idList)
         {
-            var result = await _projectEntityService.UpdateProjectEntityPriority(idList);
-            return Ok(result);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+
+            //PROJECT_MANAGER
+            if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
+            {
+                var result = await _projectEntityService.UpdateProjectEntityPriority(idList, currentUser);
+                return Ok(result);
+            }
+            return StatusCode((int)HttpStatusCode.Forbidden, "Only user with role PROJECT_MANAGER can perform this action!!!");
         }
 
+        //DELETE
         [HttpDelete]
         [Route("{id}")]
         [Authorize(Roles = "PROJECT_MANAGER")]
         public async Task<IActionResult> DeleteProjectEntity(Guid id)
         {
-            var result = await _projectEntityService.DeleteProjectEntityById(id);
-            return Ok(result);
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
+
+            //PROJECT_MANAGER
+            if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
+            {
+                var result = await _projectEntityService.DeleteProjectEntityById(id);
+                return Ok(result);
+            }
+            return StatusCode((int)HttpStatusCode.Forbidden, "Only user with role PROJECT_MANAGER can perform this action!!!");
         }
 
         private async Task<ThisUserObj> GetThisUserInfo(HttpContext? httpContext)
